pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xphys"
version = "0.1.0"
authors = [{name = "Gia Jadick"}]
description = "X-ray physics utils: material-dependent scatter coefficients and tomography library."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = ["einops >= 0.6.0", "flax >= 0.6.3", "chex>=0.1.5", "optax >=0.1.4", "jaxopt>=0.8"]

[project.optional-dependencies]
sci = ["numpy>=1.26", "matplotlib>=3.8", "scipy>=1.10.0", "jupyter", "ipykernel"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
ignore = ["F401", "F403"] #ignore unused imports errors
